Sorry
Sorry if this removes your work, but I have been working on this a bit and this is all of my work so far. I haven't been doing much, but still. I decided to update this with the version I have. My version is far from complete though :) Also, I couldn't commit the md5 database as it is too big. Sorry about that.